+++ openrouteservice/src/test/java/org/heigit/ors/routing/graphhopper/extensions/storages/builders/WheelchairGraphStorageBuilderTest.java
@@ -27,12 +27,13 @@ public class WheelchairGraphStorageBuilderTest {
     @Test
     public void TestProcessSeparateWay() {
         ReaderWay way = new ReaderWay(1);
+        way.setTag("highway", "residential");
         way.setTag("width", "0.5");
         way.setTag("incline", "2");
         way.setTag("kerb:height", "0.03");
         way.setTag("smoothness", "good");
         way.setTag("surface", "asphalt");
-        way.setTag("tracktype", "grade4");
+
 
         builder.processWay(way);
 
@@ -42,7 +43,30 @@ public class WheelchairGraphStorageBuilderTest {
         Assert.assertEquals(3, attrs.getSlopedKerbHeight());
         Assert.assertEquals(2, attrs.getSmoothnessType());
         Assert.assertEquals(2, attrs.getSurfaceType());
-        Assert.assertEquals(4, attrs.getTrackType());
+        Assert.assertFalse(attrs.isSurfaceQualityKnown());
+        Assert.assertFalse(attrs.isPedestrianised());
+    }
+
+
+    @Test
+    public void TestPedestrianisedWay() {
+        ReaderWay way = new ReaderWay(1);
+        way.setTag("highway", "track");
+        way.setTag("width", "0.5");
+        way.setTag("incline", "2");
+        way.setTag("tracktype", "grade1");
+        way.setTag("surface", "asphalt");
+
+
+        builder.processWay(way);
+
+        WheelchairAttributes attrs = builder.getStoredAttributes(WheelchairGraphStorageBuilder.Side.NONE);
+        Assert.assertEquals(50, attrs.getWidth());
+        Assert.assertEquals(2, attrs.getIncline());
+        Assert.assertEquals(2, attrs.getSurfaceType());
+        Assert.assertEquals(1, attrs.getTrackType());
+        Assert.assertTrue(attrs.isSurfaceQualityKnown());
+        Assert.assertTrue(attrs.isPedestrianised());
     }
 
     @Test
@@ -95,9 +119,21 @@ public class WheelchairGraphStorageBuilderTest {
         correctWheelchairAttributes.setAttribute(WheelchairAttributes.Attribute.TRACK, 2);
         correctWheelchairAttributes.setAttribute(WheelchairAttributes.Attribute.SMOOTHNESS, 3);
         correctWheelchairAttributes.setAttribute(WheelchairAttributes.Attribute.SURFACE, 4);
+        Assert.assertFalse(correctWheelchairAttributes.isSurfaceQualityKnown());
+        Assert.assertFalse(correctWheelchairAttributes.isPedestrianised());
+
         builder.processWay(way);
+        WheelchairAttributes left_attrs = builder.getStoredAttributes(WheelchairGraphStorageBuilder.Side.LEFT);
+        Assert.assertTrue(left_attrs.isSurfaceQualityKnown());
+        Assert.assertTrue(left_attrs.isPedestrianised());
+
+        WheelchairAttributes right_attrs = builder.getStoredAttributes(WheelchairGraphStorageBuilder.Side.RIGHT);
+        Assert.assertTrue(right_attrs.isSurfaceQualityKnown());
+        Assert.assertTrue(right_attrs.isPedestrianised());
 
         WheelchairAttributes attrs = builder.combineAttributesOfWayWhenBothSidesPresent(new WheelchairAttributes());
+        Assert.assertFalse(attrs.isSurfaceQualityKnown());
+        Assert.assertFalse(attrs.isPedestrianised());
 
         Assert.assertEquals(wheelchairAttributesAsString(correctWheelchairAttributes), wheelchairAttributesAsString(attrs));
     }
@@ -165,6 +201,8 @@ public class WheelchairGraphStorageBuilderTest {
         Assert.assertEquals(2, attrs.getSmoothnessType());
         Assert.assertEquals(2, attrs.getSurfaceType());
         Assert.assertEquals(4, attrs.getTrackType());
+        Assert.assertTrue(attrs.isPedestrianised());
+        Assert.assertTrue(attrs.isSurfaceQualityKnown());
     }
 
     private String wheelchairAttributesAsString(WheelchairAttributes attrs) {